backend/lifemanager/urls.py
"""
URL configuration for lifemanager project.
"""
from django.contrib import admin
from django.urls import path, include
from django.conf import settings
from django.conf.urls.static import static
from rest_framework_simplejwt.views import (
    TokenObtainPairView,
    TokenRefreshView,
)

urlpatterns = [
    path('admin/', admin.site.urls),
    path('api/auth/token/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
    path('api/auth/token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
    path('api/auth/', include('apps.auth.urls')),
    path('api/tasks/', include('apps.tasks.urls')),
    path('api/finances/', include('apps.finances.urls')),
    path('api/documents/', include('apps.documents.urls')),
    path('api/reminders/', include('apps.reminders.urls')),
    path('api/notes/', include('apps.notes.urls')),
]

if settings.DEBUG:
    urlpatterns += static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
    urlpatterns += static(settings.STATIC_URL, document_root=settings.STATIC_ROOT)
