backend/requirements.txt
Django==4.2.11
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.2
psycopg2-binary==2.9.9
python-decouple==3.8
celery==5.3.4
redis==5.0.1
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
django-storages==1.14.2
boto3==1.34.32
pillow==10.1.0
django-filter==23.5
python-dateutil==2.8.2
pytest==7.4.3
pytest-django==4.7.0
pytest-asyncio==0.21.1
factory-boy==3.3.0
faker==21.0.0
freezegun==1.3.0
gunicorn==21.2.0
whitenoise==6.6.0
